Bomb Bug Fix, Add Bomb Bug

+++ BomberMan/GameMap.h
@@ -13,8 +13,8 @@ public:
 	static void Update();
 	static void Render();
 
-	static void SetBlock(const int x, const int y);
-	static void RemoveBlock(const int x, const int y);
+	static void SetBlock(const double x, const double y);
+	static void RemoveBlock(const double x, const double y);
 
 	static const bool IsMovePoint(const double px, const double py);
 };
